fix: user data async & cannot read property error
