+++ src/Components/PostReplyModal/index.jsx
@@ -5,8 +5,7 @@ import { ReactComponent as Talk } from '../../Assets/icon/talk.svg';
 import { getOneTweet } from '../../API/getOneTweet';
 import { useAuth } from '../../Context/AuthContext';
 
-export default function PostReplyModal({ className, tweetId, handleRender }) {
-  const signinUser = useAuth().currentMember;
+export default function PostReplyModal({ className, tweetId, signinUser, handleRender }) {
   const [tweetInfo, setTweetInfo] = useState({});
   const [showModal, setShowModal] = useState(false);
   const handleOnClick = async e => {
